// packages/metafox/user/src/types.tsx
export type SharingPrivacyItem = {
  module_id: string;
  phrase: string;
  var_name: string;
  value: number;
  custom_id: string;
  options: { label: string; value: number }[];
};